+++ website/src/utils/codeeditor/typescript/index.ts
@@ -23,7 +23,7 @@ export const project: Project = {
     react: '^18.2.0',
     'react-dom': '^18.2.0',
     'react-router-dom': '^6.26.1',
-    '@multisynq/react-together': '^0.3.0',
+    'react-together': '^0.3.0',
     '@types/react': '^18.2.71',
     '@types/react-dom': '^18.2.22',
     primereact: '^10.6.6',